requirements.txt
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
gallery_dl==1.27.0
idna==3.7
mutagen==1.47.0
pyaes==1.6.1
pycryptodomex==3.20.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
TgCrypto==1.2.5
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.5.27
